player class in another file
